Parse set_box assignment, and do parsing generally in a much more robust way
Instead of looking for not-parse parse not-parse pattern, we keep
parsing until we get a syntax error. We ignore no-more-token errors,
because this means we can get more input and have a valid command.
The previous approach did not work if a command can always be
extended in a parse-able manner.